+++ src/main/java/com/elmakers/mine/bukkit/plugins/spells/builtin/WolfSpell.java
@@ -5,7 +5,6 @@ import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.craftbukkit.entity.CraftWolf;
 import org.bukkit.entity.CreatureType;
-import org.bukkit.entity.Entity;
 
 import com.elmakers.mine.bukkit.plugins.spells.Spell;
 import com.elmakers.mine.bukkit.plugins.spells.Target;
@@ -40,14 +39,10 @@ public class WolfSpell extends Spell
 	public boolean onCast(String[] parameters)
 	{
 	    this.targetEntity(CraftWolf.class);
-		Target target = getTarget();
-		if (target.hasTarget() && target.isEntity())
+		Target target = getTargetEntity();
+		if (target == null || !target.hasTarget() || !(target.getEntity() instanceof CraftWolf))
 		{
-    		Entity targetWolf = target.getEntity();
-    		if (!(targetWolf instanceof CraftWolf))
-    		{
-    			return newWolf();
-    		}
+			return newWolf();
 		}
 		
 		CraftWolf wolfie = (CraftWolf)target.getEntity();